Moved the KBV questions pages to a folder
