+++ tests/audit-tests/src/test/java/pages/KBV/FindYourAddressPage.java
@@ -1,4 +1,4 @@
-package pages;
+package pages.KBV;
 
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
