+++ tests/audit-tests/src/test/java/pages/KBV/HowMuchDoYouHaveLeftToPayOnYourMortgagePage.java
@@ -1,4 +1,4 @@
-package pages;
+package pages.KBV;
 
 //import gov.di_ipv_kbv.utilities.Driver;
 
@@ -7,15 +7,18 @@ import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
 import utilities.Driver;
 
-public class WhatIsTheNameOfYourLoanProviderPage {
-    public WhatIsTheNameOfYourLoanProviderPage() {
+public class HowMuchDoYouHaveLeftToPayOnYourMortgagePage {
+    public HowMuchDoYouHaveLeftToPayOnYourMortgagePage() {
         PageFactory.initElements(Driver.get(), this);
     }
 
-    @FindBy(id = "Q00033-TSBBANKPLC-label")
-    public WebElement TSBBANKPLC;
+    @FindBy(id = "Q00015-OVER35000UPTO60000-label")
+    public WebElement OVER35000UPTO60000;
 
-    @FindBy(id = "Q00033-NONEOFTHEABOVEDOESNOTAPPLY-label")
+    @FindBy(id = "Q00015-UPTO60000-label")
+    public WebElement UPTO60000;
+
+    @FindBy(id = "Q00015-NONEOFTHEABOVEDOESNOTAPPLY-label")
     public WebElement NONEOFTHEABOVEDOESNOTAPPLY;
 
     @FindBy(id = "continue")
