+++ tests/audit-tests/src/test/java/pages/KBV/HowMuchOfYourLoanDoYouHaveLeftToPayBackPage.java
@@ -1,4 +1,4 @@
-package pages;
+package pages.KBV;
 
 //import gov.di_ipv_kbv.utilities.Driver;
 
@@ -7,18 +7,18 @@ import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
 import utilities.Driver;
 
-public class HowMuchOfYourLoanDoYouPayBackEveryMonthPage {
-    public HowMuchOfYourLoanDoYouPayBackEveryMonthPage() {
+public class HowMuchOfYourLoanDoYouHaveLeftToPayBackPage {
+    public HowMuchOfYourLoanDoYouHaveLeftToPayBackPage() {
         PageFactory.initElements(Driver.get(), this);
     }
 
-    @FindBy(id = "Q00042-OVER550UPTO600-label")
-    public WebElement OVER550UPTO600;
+    @FindBy(id = "Q00039-UPTO6750-label")
+    public WebElement UPTO6750;
 
-    @FindBy(id = "Q00042-UPTO600-label")
-    public WebElement UPTO600;
+    @FindBy(id = "Q00039-OVER6500UPTO6750-label")
+    public WebElement OVER6500UPTO6750;
 
-    @FindBy(id = "Q00042-NONEOFTHEABOVEDOESNOTAPPLY-label")
+    @FindBy(id = "Q00039-NONEOFTHEABOVEDOESNOTAPPLY-label")
     public WebElement NONEOFTHEABOVEDOESNOTAPPLY;
 
     @FindBy(id = "continue")
