+++ tests/audit-tests/src/test/java/pages/KBV/HowMuchOfYourLoanDoYouPayBackEveryMonthPage.java
@@ -1,4 +1,4 @@
-package pages;
+package pages.KBV;
 
 //import gov.di_ipv_kbv.utilities.Driver;
 
@@ -7,15 +7,18 @@ import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
 import utilities.Driver;
 
-public class WhatAreTheFirstTwoLettersOfTheForenameOfTheOtherPersonOnYourMortgagePage {
-    public WhatAreTheFirstTwoLettersOfTheForenameOfTheOtherPersonOnYourMortgagePage() {
+public class HowMuchOfYourLoanDoYouPayBackEveryMonthPage {
+    public HowMuchOfYourLoanDoYouPayBackEveryMonthPage() {
         PageFactory.initElements(Driver.get(), this);
     }
 
-    @FindBy(xpath = "//label[@id='Q00019-KA-label']")
-    public WebElement KA;
+    @FindBy(id = "Q00042-OVER550UPTO600-label")
+    public WebElement OVER550UPTO600;
 
-    @FindBy(id = "Q00019-NONEOFTHEABOVEDOESNOTAPPLY-label")
+    @FindBy(id = "Q00042-UPTO600-label")
+    public WebElement UPTO600;
+
+    @FindBy(id = "Q00042-NONEOFTHEABOVEDOESNOTAPPLY-label")
     public WebElement NONEOFTHEABOVEDOESNOTAPPLY;
 
     @FindBy(id = "continue")
