+++ tests/audit-tests/src/test/java/pages/KBV/WhatAreTheFirstTwoLettersOfTheForenameOfTheOtherPersonOnYourMortgagePage.java
@@ -1,4 +1,4 @@
-package pages;
+package pages.KBV;
 
 //import gov.di_ipv_kbv.utilities.Driver;
 
@@ -7,18 +7,15 @@ import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
 import utilities.Driver;
 
-public class HowMuchOfYourLoanDoYouHaveLeftToPayBackPage {
-    public HowMuchOfYourLoanDoYouHaveLeftToPayBackPage() {
+public class WhatAreTheFirstTwoLettersOfTheForenameOfTheOtherPersonOnYourMortgagePage {
+    public WhatAreTheFirstTwoLettersOfTheForenameOfTheOtherPersonOnYourMortgagePage() {
         PageFactory.initElements(Driver.get(), this);
     }
 
-    @FindBy(id = "Q00039-UPTO6750-label")
-    public WebElement UPTO6750;
+    @FindBy(xpath = "//label[@id='Q00019-KA-label']")
+    public WebElement KA;
 
-    @FindBy(id = "Q00039-OVER6500UPTO6750-label")
-    public WebElement OVER6500UPTO6750;
-
-    @FindBy(id = "Q00039-NONEOFTHEABOVEDOESNOTAPPLY-label")
+    @FindBy(id = "Q00019-NONEOFTHEABOVEDOESNOTAPPLY-label")
     public WebElement NONEOFTHEABOVEDOESNOTAPPLY;
 
     @FindBy(id = "continue")
